new/brain_py/reversal/requirements.txt
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
onnx>=1.14.0
onnxmltools>=1.12.0
